Change names - remove debug comments
