+++ sample-apps/radiology/lib/infers/localization_spine.py
@@ -30,7 +30,7 @@ from monailabel.interfaces.tasks.infer import InferTask, InferType
 from monailabel.transform.post import Restored
 
 
-class SpineLoc(InferTask):
+class LocalizationSpine(InferTask):
     """
     This provides Inference Engine for pre-trained spine localization (UNet) model.
     """
@@ -73,15 +73,8 @@ class SpineLoc(InferTask):
         return SlidingWindowInferer(roi_size=self.roi_size)
 
     def post_transforms(self, data=None) -> Sequence[Callable]:
-        largest_cc = False if not data else data.get("largest_cc", False)
-        applied_labels = list(self.labels.values()) if isinstance(self.labels, dict) else self.labels
-        t = [
+        return [
             EnsureTyped(keys="pred", device=data.get("device") if data else None),
-            Activationsd(keys="pred", softmax=True),
-            AsDiscreted(keys="pred", argmax=True),
-            ToNumpyd(keys="pred"),
+            Activationsd(keys="pred", softmax=True), AsDiscreted(keys="pred", argmax=True), ToNumpyd(keys="pred"),
+            Restored(keys="pred", ref_image="image")
         ]
-        if largest_cc:
-            t.append(KeepLargestConnectedComponentd(keys="pred", applied_labels=applied_labels))
-        t.append(Restored(keys="pred", ref_image="image"))
-        return t
