+++ sample-apps/radiology/lib/trainers/__init__.py
@@ -11,8 +11,8 @@
 
 from .deepedit import DeepEdit
 from .deepgrow import Deepgrow
+from .localization_spine import LocalizationSpine
+from .localization_vertebra import LocalizationVertebra
 from .segmentation import Segmentation
 from .segmentation_spleen import SegmentationSpleen
-from .spine_loc import SpineLoc
-from .ver_loc import VerLoc
-from .ver_seg import VerSeg
+from .segmentation_vertebra import SegmentationVertebra
