+++ sample-apps/radiology/lib/transforms/transforms.py
@@ -129,38 +129,10 @@ class GetCentroidAndCropd(MapTransform):
                 d[key][d[key] != first_label["label"]] = 0
                 d[key][d[key] > 0] = 1
 
-                # Plotting
-                # from matplotlib.pyplot import imshow, show, close
-                # imshow(d[key][0,:,:,int(d[key].shape[-1]/2)])
-                # show()
-                # close()
             elif key == "image":
                 d[key] = cropper(d[key])
             else:
                 print("This transform only applies to the label or image")
-
-        # For debugging purposes
-        # canvas_img = np.zeros(d["original_size"], dtype=np.float32)
-        # canvas_label = np.zeros(d["original_size"], dtype=np.float32)
-        #
-        # canvas_img[
-        #     cropper.slices[-3].start : cropper.slices[-3].stop,
-        #     cropper.slices[-2].start : cropper.slices[-2].stop,
-        #     cropper.slices[-1].start : cropper.slices[-1].stop,
-        # ] = d["image"]
-        #
-        # canvas_label[
-        #     cropper.slices[-3].start : cropper.slices[-3].stop,
-        #     cropper.slices[-2].start : cropper.slices[-2].stop,
-        #     cropper.slices[-1].start : cropper.slices[-1].stop,
-        # ] = d["label"]
-        #
-        # d["image"] = canvas_img
-        # d["label"] = canvas_label
-        #
-        # SaveImaged(keys="image", output_postfix="", output_dir="/home/andres/Downloads", separate_folder=False)(d)
-        # SaveImaged(keys="label", output_postfix="seg", output_dir="/home/andres/Downloads", separate_folder=False)(d)
-
         return d
 
 
@@ -306,52 +278,7 @@ class VertHeatMap(MapTransform):
 
             data[k] = out
 
-            # SaveImaged(keys="label", output_postfix="", output_dir="/home/andres/Downloads", separate_folder=False)(
-            #     data
-            # )
-
         return data
-
-
-# class VertHeatMap(MapTransform):
-#     def __init__(self, keys, gamma=1000.0, label_names=None):
-#         super().__init__(keys)
-#         self.label_names = label_names
-#         self.gamma = gamma
-#
-#     def __call__(self, data):
-#
-#         for k in self.keys:
-#
-#             out = np.zeros(
-#                 (len(self.label_names) + 1, data[k].shape[-3], data[k].shape[-2], data[k].shape[-1]), dtype=np.float32
-#             )
-#
-#             # loop over all segmentation classes
-#             for label_num in np.unique(data[k]):
-#                 # skip background
-#                 if label_num == 0:
-#                     continue
-#                 # get CoM for given segmentation class
-#                 centre = [np.average(indices).astype(int) for indices in np.where(data[k][0] == label_num)]
-#                 centre.insert(0, label_num)
-#                 out[int(centre[-4]), int(centre[-3]), int(centre[-2]), int(centre[-1])] = 1.0
-#                 sigma = 1.6 + (label_num - 1.0) * 0.1
-#                 # Gaussian smooth
-#                 out[int(label_num)] = GaussianSmooth(sigma)(out[int(label_num)])
-#                 # # Normalize to [0,1]
-#                 out[int(label_num)] = ScaleIntensity()(out[int(label_num)])
-#                 out[int(label_num)] = out[int(label_num)] * self.gamma
-#
-#             # TO DO: Keep the centroids in the data dictionary?
-#
-#             data[k] = out
-#
-#             # SaveImaged(keys="label", output_postfix="", output_dir="/home/andres/Downloads", separate_folder=False)(
-#             #     data
-#             # )
-#
-#         return data
 
 
 class VertebraLocalizationPostProcessing(MapTransform):
@@ -382,13 +309,6 @@ class VertebraLocalizationPostProcessing(MapTransform):
                 X, Y, Z = X[0], Y[0], Z[0]
                 centroid[f"label_{l+1}"] = [X, Y, Z]
                 centroids.append(centroid)
-
+            d["pred_centroids"] = centroids
             print(centroids)
-            # d["pred_meta_dict"] = d["image_meta_dict"]
-            # SaveImaged(keys=key, output_postfix="", output_dir="/home/andres/Downloads", separate_folder=False)(d)
-            # # Plotting
-            # from matplotlib.pyplot import imshow, show, close
-            # imshow(d[key][0,:,:,int(d[key].shape[-1]/2)])
-            # show()
-            # close()
         return d
